Implementando os casos de teste

+++ src/main/java/br/edu/iftm/atividadeComplementar/repositories/AtividadeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import br.edu.iftm.atividadeComplementar.domains.Atividade;
 
-public interface AtividadeRepository extends JpaRepository<Atividade, Integer> {
+public interface AtividadeRepository extends JpaRepository<Atividade, Long> {
 
 	public List<Atividade> findByNomeContaining(@Param("nome") String nome);
